Задача спринта 22

+++ app/src/main/java/com/example/playlist_maker/music_library/domain/FavoriteTracksInteractorImpl.kt
@@ -17,7 +17,7 @@ class FavoriteTracksInteractorImpl(private val favoriteTracksRepository: Favorit
         return favoriteTracksRepository.deleteTrack(trackId = trackId)
     }
 
-    override suspend fun isFavoriteTrack(trackId: Long): Flow<Boolean> {
+    override suspend fun isFavoriteTrack(trackId: Int): Flow<Boolean> {
         return favoriteTracksRepository.isFavoriteTrack(trackId)
     }
 }
